Moved common types from index to types.ts to import in other modules in shared
