+++ loft-taxi/src/shared/types.ts
@@ -1,0 +1,11 @@
+export enum PageID {
+  LOGIN = 'login',
+  PROFILE = 'profile',
+  MAP = 'map',
+  REGISTRATION = 'registration',
+}
+
+export type PageInfo = {
+  id: PageID;
+  title: string;
+};
